remove unnecessary doc, add note about user owner
With one exception (technical difficulties w/ crates.io at https://github.com/georust/geos/issues/80) this has been completed for all our crates, so no need to clutter up the docs with steps which no longer apply to anyone.

Since this is a minor change that doesn't change the spirit of the guideline, I'm not planning on trying to solicit a bunch of feedback. Happy to hear any though - I'll leave it open for a week.